Version bump to v3.0.0-beta.36
+ Settings split and organized
